memory: remove ranges
Add support to remove a memory range by calling madvise(MADV_REMOVE).
This frees up the underlying storage for the removed pages, without
changing the total size of the used memory (although VmRss does increase
as the madvise acts like the guest touching the pages). This opens the
gate for ballooning.
The pages in the specified range back are immediately returned to the
kernel. Subsequent accesses to memory in that range return zero pages.

Signed-off-by: Alexandra Iordache <aghecen@amazon.com>